#3 Clean view and rename files
