+++ README.md
@@ -4,7 +4,7 @@
 
 #### 1. Get the latest version
 
-You can start by cloning the latest version of React Dashboard on your
+You can start by cloning the latest version of React Home on your
 local machine by running:
 
 ```shell
