+++ src/components/Layout/LayoutView.js
@@ -7,7 +7,7 @@ import Header from '../Header';
 import Sidebar from '../Sidebar';
 
 // pages
-import Dashboard from '../../pages/dashboard';
+import Home from '../../pages/home';
 import Typography from '../../pages/typography';
 import Notifications from '../../pages/notifications';
 import Maps from '../../pages/maps';
@@ -25,7 +25,7 @@ const Layout = ({ classes, isSidebarOpened, toggleSidebar }) => (
         <div className={classnames(classes.content, { [classes.contentShift]: isSidebarOpened })}>
           <div className={classes.fakeToolbar} />
           <Switch>
-            <Route path="/app/dashboard" component={Dashboard} />
+            <Route path="/app/home" component={Home} />
             <Route path="/app/typography" component={Typography} />
             <Route path="/app/tables" component={Tables} />
             <Route path="/app/notifications" component={Notifications} />
