+++ src/components/Sidebar/SidebarView.js
@@ -21,8 +21,8 @@ import SidebarLink from './components/SidebarLink/SidebarLinkContainer';
 import Dot from './components/Dot';
 
 const structure = [
-  { id: 0, label: 'Dashboard', link: '/app/dashboard', icon: <HomeIcon /> },
-  { id: 1, label: 'Typography', link: '/app/typography', icon: <TypographyIcon /> },
+  { id: 0, label: 'Home', link: '/app/home', icon: <HomeIcon /> }
+  /*,{ id: 1, label: 'Typography', link: '/app/typography', icon: <TypographyIcon /> },
   { id: 2, label: 'Tables', link: '/app/tables', icon: <TableIcon /> },
   { id: 3, label: 'Notifications', link: '/app/notifications', icon: <NotificationsIcon />},
   {
@@ -45,7 +45,7 @@ const structure = [
   { id: 11, type: 'title', label: 'PROJECTS' },
   { id: 12, label: 'My recent', link: '', icon: <Dot size="large" color="warning" /> },
   { id: 13, label: 'Starred', link: '', icon: <Dot size="large" color="primary" /> },
-  { id: 14, label: 'Background', link: '', icon: <Dot size="large" color="secondary" /> },
+  { id: 14, label: 'Background', link: '', icon: <Dot size="large" color="secondary" /> },*/
 ];
 
 const SidebarView = ({ classes, theme, toggleSidebar, isSidebarOpened, isPermanent, location }) => {
