--- src/pages/dashboard/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "Dashboard",
-  "version": "1.0.0",
-  "private": true,
-  "main": "DashboardContainer.js"
-}
