+++ src/pages/home/components/BigStat/BigStat.js
