+++ src/pages/home/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "Home",
+  "version": "1.0.0",
+  "private": true,
+  "main": "HomeContainer.js"
+}
